% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_matrix.R
\docType{data}
\name{example_matrix}
\alias{example_matrix}
\title{Example Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 2 columns.
}
\usage{
example_matrix
}
\description{
\eqn{5 \times 2} matrix intended for use as an example in dmvvg.
}
\author{
Samuel Soon
}
\keyword{datasets}
